# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Circle-gated stock & options screener
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LogoutResult {
  success: boolean;
}

/**
 * Authorization mode
 */
export type AuthStatusMode = typeof AuthStatusMode[keyof typeof AuthStatusMode];


export const AuthStatusMode = {
  mock: 'mock',
  live: 'live',
} as const;

export interface CircleUser {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface AuthStatus {
  /** Whether the user has a valid Circle identity */
  authenticated: boolean;
  /** Whether the user has access to the required Space Group */
  authorized: boolean;
  /** Authorization mode */
  mode: AuthStatusMode;
  user?: CircleUser;
  /**
     * URL to initiate login (present when not authenticated)
     * @nullable
     */
  loginUrl?: string | null;
}

export type MarketStatusState = typeof MarketStatusState[keyof typeof MarketStatusState];


export const MarketStatusState = {
  open: 'open',
  closed: 'closed',
  pre_market: 'pre_market',
  after_hours: 'after_hours',
} as const;

export interface MarketStatus {
  state: MarketStatusState;
  label: string;
  description: string;
  timestamp: string;
  /** @nullable */
  nextOpen?: string | null;
  /** @nullable */
  nextClose?: string | null;
}

export interface FilterResult {
  name: string;
  passed: boolean;
  calculatedValue: string;
  threshold: string;
  explanation: string;
}

export type StockResultStatus = typeof StockResultStatus[keyof typeof StockResultStatus];


export const StockResultStatus = {
  qualified: 'qualified',
  not_qualified: 'not_qualified',
} as const;

export interface StockResult {
  symbol: string;
  company: string;
  price: number;
  dailyChangePercent: number;
  volume: number;
  avgVolume: number;
  marketCap: number;
  impliedVolatility: number;
  optionsVolume: number;
  openInterest: number;
  filterScore: number;
  status: StockResultStatus;
  qualified: boolean;
  filterResults: FilterResult[];
}

export interface ScanResult {
  stocks: StockResult[];
  totalScanned: number;
  totalQualified: number;
  scanTime: string;
  dataAsOf: string;
}

export type ScannerStateStatus = typeof ScannerStateStatus[keyof typeof ScannerStateStatus];


export const ScannerStateStatus = {
  idle: 'idle',
  running: 'running',
  complete: 'complete',
  error: 'error',
} as const;

export interface ScannerState {
  hasResults: boolean;
  status: ScannerStateStatus;
  /** Most recent scan result, or null if no scan has been run yet */
  lastScan?: ScanResult | null;
  /** @nullable */
  lastScanTime?: string | null;
}

export interface OptionsContract {
  strike: number;
  expiration: string;
  callBid: number;
  callAsk: number;
  callVolume: number;
  callOI: number;
  callIV: number;
  callDelta: number;
  callGamma: number;
  callTheta: number;
  callVega: number;
  putBid: number;
  putAsk: number;
  putVolume: number;
  putOI: number;
  putIV: number;
  putDelta: number;
  putGamma: number;
  putTheta: number;
  putVega: number;
}

export interface OptionsChain {
  symbol: string;
  expirations: string[];
  contracts: OptionsContract[];
}

export interface StockDetail {
  stock: StockResult;
  optionsChain: OptionsChain;
}

export type InitiateLoginParams = {
/**
 * Dev mock only: which scenario to simulate
 */
scenario?: InitiateLoginScenario;
};

export type InitiateLoginScenario = typeof InitiateLoginScenario[keyof typeof InitiateLoginScenario];


export const InitiateLoginScenario = {
  authorized: 'authorized',
  unauthorized: 'unauthorized',
  anonymous: 'anonymous',
} as const;

export type HandleAuthCallbackParams = {
code?: string;
state?: string;
};