# lib/api-zod/src/generated/types/circleUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Circle-gated stock & options screener
 * OpenAPI spec version: 0.1.0
 */

export interface CircleUser {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
}