# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Circle-gated stock & options screener
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}