# lib/api-zod/src/generated/types/optionsChain.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Circle-gated stock & options screener
 * OpenAPI spec version: 0.1.0
 */
import type { OptionsContract } from './optionsContract';

export interface OptionsChain {
  symbol: string;
  expirations: Date[];
  contracts: OptionsContract[];
}